Improve form validation for file type fields
The form validation of the fields with type file was triggered when
the field lost the focus via the 'onBlur' JS event (see commit 20d1bab).

In order to improve the user experience, the validation of the fields
with type file is from now on triggered by the 'onChange' JS event.
This way, once the user chooses the file, the "No file provided"
error message is immediately removed from the field.